+++ examples/chat/createChatCompletionStream.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2023-present, Sascha Greuel and Contributors
+ * Copyright (c) 2024, Sascha Greuel and Contributors
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -18,20 +18,23 @@
 
 require_once __DIR__ . '/../PerplexityAIFactory.php';
 
-// Call the createChatCompletion method.
+// Call the createChatCompletion method with specific options for streaming support.
 PerplexityAIFactory::request(
     'createChatCompletion',
+    [],
     [
         'model' => 'llama-3.1-sonar-small-128k-online',
         'messages' => [
             [
-                'role' => 'system',
-                'content' => 'Be precise and concise.',
-            ],
-            [
                 'role' => 'user',
-                'content' => 'How many stars are there in our galaxy?',
+                'content' => 'Tell me a story about a brave knight.',
             ],
         ],
-    ]
+        'stream' => true,
+    ],
+    static function ($data) {
+        if (isset($data['choices'][0]['delta']['content'])) {
+            echo $data['choices'][0]['delta']['content'];
+        }
+    }
 );